+++ src/modules/popup/details/Images.tsx
@@ -1,9 +1,11 @@
-import {type ImageData} from '_scripts/imagesExtractor'
+import {types, ImageData, ImageType} from '_scripts/imagesExtractor'
 import {useEffect, useState} from 'react'
 
 import Layout from '~/Global/Layout'
+import Controls, {ControlsTabItem, CONTROLS_GAP} from '~/UI/Controls'
 import Gallery from '~~popup/images/Gallery'
-import {ControlsBack} from '~/UI/Controls'
+
+export type ImageMode = Exclude<ImageType, 'bg-image'> | 'all'
 
 export type ImagesResponse = {
   images: ImageData[]
@@ -11,6 +13,7 @@ export type ImagesResponse = {
 
 export function Images() {
   const [imagesData, setImagesData] = useState<ImagesResponse>()
+  const [imageFormat, setImageFormat] = useState<ImageMode>('all')
 
   useEffect(() => {
     chrome.tabs.query({active: true, currentWindow: true}, (tabs) => {
@@ -22,11 +25,19 @@ export function Images() {
     })
   }, [])
 
+  const handleFormatChange = (format: ImageMode) => {
+    setImageFormat(format)
+  }
+
   return (
     <Layout className="space-y-3.5">
-      <ControlsBack />
+      <Controls>
+        {['all', ...types.filter((type) => type !== 'bg-image')].map((type) => (
+          <ControlsTabItem item={type} isSelected={type === imageFormat} onClick={() => handleFormatChange(type as ImageMode)} key={type} />
+        ))}
+      </Controls>
 
-      <Gallery data={imagesData} />
+      <Gallery data={imagesData} format={imageFormat} className={CONTROLS_GAP} />
     </Layout>
   )
 }